requirements.txt
h5py
numpy
pillow
pytesseract
requests
tqdm